sseder and migration change
sseder and migration change

+++ database/migrations/2024_01_01_000064_remove_role_from_users_table.php
@@ -9,7 +9,9 @@ class RemoveRoleFromUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
